feat: 모임신청가이드 및 svg 되돌리기
