+++ pages/list/index.tsx
@@ -13,8 +13,8 @@ import { useRouter } from 'next/router';
 import { styled } from 'stitches.config';
 import CrewTab from '@components/CrewTab';
 
-import ArrowRightCircleIcon from '@assets/svg/arrow_right_circle.svg';
 import FloatingButton from '@components/FloatingButton';
+import GuideButton from '@components/GuideButton';
 
 const Home: NextPage = () => {
   const router = useRouter();
@@ -35,14 +35,7 @@ const Home: NextPage = () => {
       <div>
         {/*크루 탭 - 홈, 전체 모임, 내모임, 모임 신청 가이드 */}
         <CrewTab>
-          <SGuideButton
-            target="_blank"
-            href="https://www.notion.so/sopt-makers/eec46a4562ec48f0b0220153bb6ea68e"
-            rel="noreferrer noopener"
-          >
-            모임 신청 가이드
-            <ArrowRightCircleIcon />
-          </SGuideButton>
+          <GuideButton />
         </CrewTab>
 
         {/*필터 - 드롭다운, 모임 검색*/}
@@ -88,25 +81,5 @@ const SFilterWrapper = styled('div', {
   },
   '@mobile': {
     mb: '$28',
-  },
-});
-
-const SGuideButton = styled('a', {
-  flexType: 'verticalCenter',
-  gap: '$8',
-  color: '$gray10',
-  padding: '$8 $6 0 0',
-  fontAg: '18_semibold_100',
-
-  '@tablet': {
-    padding: '0 $12 0 0',
-    fontAg: '14_semibold_100',
-  },
-
-  path: {
-    stroke: '$gray10',
-  },
-  '@media (max-width: 359px)': {
-    display: 'none',
   },
 });
